backend/requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
dotenv==0.9.9
Flask==3.0.3
Flask-Cors==4.0.1
google-auth==2.34.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
urllib3==2.5.0
Werkzeug==3.1.3
